Add README update and CONTRIBUTING
Currently, the README is useless and there's no guidance for contributing to this repo. I don't anticipate there being anyone else contributing but I figured it could be good to just get my thoughts on things into a doc somewhere, if for no other reason than the mental exercise. Anyway, this change adds instructions for running the server in its current form and adds my thoughts on commits in the new CONTRIBUTING.md. Yey.