Add Dockerfile and Makefile
This adds a Dockerfile and complimentary Makefile.

It's 2021, people, and the only way to ship anything is in a Docker container, so this adds a basic Dockerfile to build and run this go server in. It's very much based on the official golang's dockerhub example. In fact it's almost an exact copy. This builds and runs the server in one fell swoop inside the container. To complement the Dockerfile, I created a Makefile to easily run the correct Docker commands where ever and whoever you are.