Create first surveys API with gin
Currently, this server is just a boring old server that does...not much. This change implements a small surveys API using the gin library and this tutorial; https://golang.org/doc/tutorial/web-service-gin . This is the next step towards the health check app. Next steps include adding a database and https.