Add docker-compose.yml for db times
Currently, this repo doesn't have a way to start up its own database. This change adds a docker-compose.yml that contains a config for postgres 14 in it so we can start developing against a database. This also adds some handy dandy make targets for the db and renames the other make targets for consistency.