Add config.yaml
Currently, the app has no way to manage different configuration options. This change creates a config.yaml file to hold those config options. I picked a file over environment variables because environment variables can be annoying. Hard to keep track of what the options are and input can be tricky if its anything but a string. And, I feel much better about keeping things in a file inside the app instead of floating around on the current system environment. Anyway, introducing config.yaml. This file will contain anything we want to customize about the app and is easily converted into a structure we define. yey.