Add GET /teams
This is version one of this feature and may be subjet to immediate change, but in the spirit of iteration and AGILE i wanted to get something up. With this change, the /teams endpoint reaches out to github with creds (specified in config.yaml) and returns the content of teams.json. Issues that I see already:
1. Do we need to use a whole oauth library to send the token'd request to github? seems like maybe no.
2. When we get teams, it's just their names and I think we talked about teams having IDs. But saving them to the db would be silly, we should just bite the bullet and handle CRUD for teams. Sigh.