Add /teams GET and POST
Welcome to the new world order where you can now add a team to the database! This change sets up what I hope will be the convention going forward. API stuff will be stored in api/resource.go, endpoints are tracked in server.go, and db stuff in db/db.go. Error handling and logging could use a polish, but other than that I think this is a great start.