Add config options for loggers
This adds some new fields to config.yaml. Namely ones for logging. GORM has a builtin logger that should be pretty useful for debugging, so users can now set the level for it in config.yaml. I'm also bringing in logrus and this adds a field for that level, too.