Add some logging for the app
This change brings in logrus and gin-logrus for consistent, custom logging in the app beyong what is automatically provided by gin and GORM. It also adds some basic logging messages to things like loading the config file and existing team db operations. The GORM logger has kind of a firehose logging situation where it's either INFO and it's *everything* or its silent/warn/error and you don't get much, so the plan is to use logrus to just simply say whether or not something worked.